requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
click==8.0.1
discord.py==1.6.0
Flask==2.0.1
Flask-JWT-Extended==4.2.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
PyJWT==2.1.0
python-dotenv==0.19.0
requests==2.25.1
SQLAlchemy==1.4.17
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==2.0.1
yarl==1.6.3
